# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: earth
    description: NASA Earth observation data
  - name: farmer
    description: Farmer mode recommendations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /earth/snapshot:
    get:
      operationId: getEarthSnapshot
      tags: [earth]
      summary: Get NASA Earth observation snapshot
      description: Returns NASA POWER time-series data, FIRMS fire detections, and status metadata for the selected point.
      parameters:
        - name: latitude
          in: query
          required: false
          schema:
            type: number
            minimum: -90
            maximum: 90
            default: 26.77
        - name: longitude
          in: query
          required: false
          schema:
            type: number
            minimum: -180
            maximum: 180
            default: 31.5
        - name: days
          in: query
          required: false
          schema:
            type: integer
            minimum: 7
            maximum: 90
            default: 30
      responses:
        "200":
          description: Earth observation snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarthSnapshot"
  /power:
    get:
      operationId: getPower
      tags: [earth]
      summary: Get compact NASA POWER weather payload
      description: Returns the latest Tahta POWER readings, or a 200 cached payload when NASA POWER is unavailable.
      parameters:
        - name: latitude
          in: query
          required: false
          schema:
            type: number
            default: 26.77
        - name: longitude
          in: query
          required: false
          schema:
            type: number
            default: 31.5
        - name: days
          in: query
          required: false
          schema:
            type: integer
            minimum: 7
            maximum: 90
            default: 30
      responses:
        "200":
          description: Latest POWER readings or cached Tahta readings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PowerCompatibility"
  /recommendations:
    get:
      operationId: getRecommendations
      tags: [farmer]
      summary: Get live farmer recommendations
      description: Calculates Arabic farmer advice from the latest NASA POWER weather and soil-wetness observations for Tahta.
      parameters:
        - name: latitude
          in: query
          required: false
          schema:
            type: number
            minimum: -90
            maximum: 90
            default: 26.77
        - name: longitude
          in: query
          required: false
          schema:
            type: number
            minimum: -180
            maximum: 180
            default: 31.5
      responses:
        "200":
          description: Farmer recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendations"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PowerCompatibility:
      type: object
      required: [temp, humidity, precip, wind, soil_moisture, last_updated, status]
      properties:
        temp: { type: number }
        humidity: { type: number }
        precip: { type: number }
        wind: { type: number }
        soil_moisture: { type: number }
        last_updated: { type: string }
        status:
          type: string
          enum: [live, cached]
    EarthSnapshot:
      type: object
      required: [location, fetchedAt, period, power, fires, layers, dataStatus]
      properties:
        location:
          $ref: "#/components/schemas/EarthLocation"
        fetchedAt:
          type: string
          format: date-time
        period:
          $ref: "#/components/schemas/EarthPeriod"
        power:
          $ref: "#/components/schemas/PowerSeries"
        fires:
          $ref: "#/components/schemas/FireSummary"
        layers:
          type: array
          items:
            $ref: "#/components/schemas/EarthLayer"
        dataStatus:
          type: string
          enum: [live, cached, partial]
    EarthLocation:
      type: object
      required: [name, latitude, longitude]
      properties:
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
    EarthPeriod:
      type: object
      required: [start, end, days]
      properties:
        start:
          type: string
          format: date
        end:
          type: string
          format: date
        days:
          type: integer
    PowerSeries:
      type: object
      required: [dates, temperature, precipitation, humidity, windSpeed, solarRadiation, soilMoisture]
      properties:
        dates:
          type: array
          items: { type: string, format: date }
        temperature:
          type: array
          items: { type: number }
        precipitation:
          type: array
          items: { type: number }
        humidity:
          type: array
          items: { type: number }
        windSpeed:
          type: array
          items: { type: number }
        solarRadiation:
          type: array
          items: { type: number }
        soilMoisture:
          type: array
          items: { type: number }
        current:
          $ref: "#/components/schemas/PowerCurrent"
    PowerCurrent:
      type: object
      required: [temperature, precipitation, humidity, windSpeed, solarRadiation, soilMoisture]
      properties:
        temperature: { type: number }
        precipitation: { type: number }
        humidity: { type: number }
        windSpeed: { type: number }
        solarRadiation: { type: number }
        soilMoisture: { type: number }
    FireSummary:
      type: object
      required: [count, detections, source, status]
      properties:
        count:
          type: integer
        source:
          type: string
        status:
          type: string
          enum: [live, cached]
        detections:
          type: array
          items:
            $ref: "#/components/schemas/FireDetection"
    FireDetection:
      type: object
      required: [latitude, longitude, brightness, confidence, satellite, acquiredAt]
      properties:
        latitude: { type: number }
        longitude: { type: number }
        brightness: { type: number }
        confidence: { type: string }
        satellite: { type: string }
        instrument: { type: ["string", "null"] }
        acquiredAt: { type: string }
    EarthLayer:
      type: object
      required: [id, name, shortName, status, color, description]
      properties:
        id: { type: string }
        name: { type: string }
        shortName: { type: string }
        status: { type: string }
        color: { type: string }
        description: { type: string }
        source: { type: string }
        latestObservation: { type: ["string", "null"] }
        observationUrl: { type: ["string", "null"], format: uri }
    Recommendations:
      type: object
      required: [fetchedAt, location, weather, advice, crop, soilMoisture, rainLast7Days, source, dataStatus]
      properties:
        fetchedAt:
          type: string
          format: date-time
        location:
          $ref: "#/components/schemas/EarthLocation"
        weather:
          $ref: "#/components/schemas/FarmerWeather"
        advice:
          type: array
          items:
            $ref: "#/components/schemas/FarmerAdvice"
        crop:
          type: string
        soilMoisture:
          type: number
        rainLast7Days:
          type: number
        source:
          type: string
        dataStatus:
          type: string
          enum: [live, cached]
    FarmerWeather:
      type: object
      required: [temperature, humidity, rain, wind]
      properties:
        temperature: { type: number }
        humidity: { type: number }
        rain: { type: number }
        wind: { type: number }
    FarmerAdvice:
      type: object
      required: [id, message, tone]
      properties:
        id: { type: string }
        message: { type: string }
        tone: { type: string }