# lib/api-zod/src/generated/types/earthLayer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EarthLayer {
  id: string;
  name: string;
  shortName: string;
  status: string;
  color: string;
  description: string;
  source?: string;
  /** @nullable */
  latestObservation?: string | null;
  /** @nullable */
  observationUrl?: string | null;
}