# lib/api-zod/src/generated/types/earthSnapshot.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EarthLayer } from './earthLayer';
import type { EarthLocation } from './earthLocation';
import type { EarthPeriod } from './earthPeriod';
import type { EarthSnapshotDataStatus } from './earthSnapshotDataStatus';
import type { FireSummary } from './fireSummary';
import type { PowerSeries } from './powerSeries';

export interface EarthSnapshot {
  location: EarthLocation;
  fetchedAt: Date;
  period: EarthPeriod;
  power: PowerSeries;
  fires: FireSummary;
  layers: EarthLayer[];
  dataStatus: EarthSnapshotDataStatus;
}