# lib/api-zod/src/generated/types/farmerAdvice.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FarmerAdvice {
  id: string;
  message: string;
  tone: string;
}