# lib/api-zod/src/generated/types/farmerWeather.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FarmerWeather {
  temperature: number;
  humidity: number;
  rain: number;
  wind: number;
}