# lib/api-zod/src/generated/types/fireSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FireDetection } from './fireDetection';
import type { FireSummaryStatus } from './fireSummaryStatus';

export interface FireSummary {
  count: number;
  source: string;
  status: FireSummaryStatus;
  detections: FireDetection[];
}