# lib/api-zod/src/generated/types/powerCompatibility.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PowerCompatibilityStatus } from './powerCompatibilityStatus';

export interface PowerCompatibility {
  temp: number;
  humidity: number;
  precip: number;
  wind: number;
  soil_moisture: number;
  last_updated: string;
  status: PowerCompatibilityStatus;
}